Moved current views and serializers under 'v1' parent folder
