+++ kpi/serializers/__init__.py
@@ -1,29 +1,29 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
 
-from .ancestor_collections import AncestorCollectionsSerializer
-from .asset import AssetListSerializer
-from .asset import AssetSerializer
-from .asset import AssetUrlListSerializer
-from .asset_file import AssetFileSerializer
-from .asset_snapshot import AssetSnapshotSerializer
-from .asset_version import AssetVersionListSerializer
-from .asset_version import AssetVersionSerializer
-from .authorized_application_user import AuthorizedApplicationUserSerializer
-from .collection import CollectionListSerializer
-from .collection import CollectionSerializer
-from .collection import CollectionChildrenSerializer
-from .deployment import DeploymentSerializer
-from .export_task import ExportTaskSerializer
-from .import_task import ImportTaskListSerializer
-from .import_task import ImportTaskSerializer
-from .object_permission import ObjectPermissionNestedSerializer
-from .object_permission import ObjectPermissionSerializer
-from .one_time_authentication_key import OneTimeAuthenticationKeySerializer
-from .sitewide_message import SitewideMessageSerializer
-from .tag import TagListSerializer
-from .tag import TagSerializer
-from .user import CreateUserSerializer
-from .user import CurrentUserSerializer
-from .user import UserSerializer
-from .user_collection_subscription import UserCollectionSubscriptionSerializer
+from .v1.ancestor_collections import AncestorCollectionsSerializer
+from .v1.asset import AssetListSerializer
+from .v1.asset import AssetSerializer
+from .v1.asset import AssetUrlListSerializer
+from .v1.asset_file import AssetFileSerializer
+from .v1.asset_snapshot import AssetSnapshotSerializer
+from .v1.asset_version import AssetVersionListSerializer
+from .v1.asset_version import AssetVersionSerializer
+from .v1.authorized_application_user import AuthorizedApplicationUserSerializer
+from .v1.collection import CollectionListSerializer
+from .v1.collection import CollectionSerializer
+from .v1.collection import CollectionChildrenSerializer
+from .v1.deployment import DeploymentSerializer
+from .v1.export_task import ExportTaskSerializer
+from .v1.import_task import ImportTaskListSerializer
+from .v1.import_task import ImportTaskSerializer
+from .v1.object_permission import ObjectPermissionNestedSerializer
+from .v1.object_permission import ObjectPermissionSerializer
+from .v1.one_time_authentication_key import OneTimeAuthenticationKeySerializer
+from .v1.sitewide_message import SitewideMessageSerializer
+from .v1.tag import TagListSerializer
+from .v1.tag import TagSerializer
+from .v1.user import CreateUserSerializer
+from .v1.user import CurrentUserSerializer
+from .v1.user import UserSerializer
+from .v1.user_collection_subscription import UserCollectionSubscriptionSerializer
