+++ kpi/serializers/v1/asset.py
