+++ kpi/serializers/v1/asset_file.py
