+++ kpi/serializers/v1/asset_snapshot.py
