+++ kpi/views/v1/asset.py
