+++ kpi/views/v1/asset_file.py
