+++ kpi/views/v1/asset_snapshot.py
