+++ kpi/views/v1/collection.py
