+++ kpi/views/v1/no_update_model.py
